矩阵求逆

+++ include/function.h
@@ -1,4 +1,4 @@
-#include "myvector.h"
+#include "vector.h"
 #pragma once
 struct NdsclaFunction
 {
@@ -6,6 +6,6 @@ struct NdsclaFunction
     double (*function)(Vector *);
 };
 typedef struct NdsclaFunction NdsclaFunction;
-float NdsclaFunctionCall(NdsclaFunction *function, Vector *x);
+double NdsclaFunctionCall(NdsclaFunction *function, Vector *x);
 NdsclaFunction *NdsclaFunctionAlloc(double (*function)(Vector *), int inputsize);
 void centralGrad(NdsclaFunction *function, double h, Vector *x0, Vector *grad);
